+++ client/src/js/djng-forms.js
@@ -396,4 +396,5 @@ djng_forms_module.directive('djngBindIf', function() {
 	};
 });
 
+
 })(window.angular);